changed the def of setvar, now it needs quotes
setvar now needs the variable that it is setting to be quoted. also some
files were missing their provide and some files had a different name on
the provide compared to the file name.

this was done because i couldn't use eldoc to see the short description
of variables in my configuration because flycheck kept complaining about
free variables.