using diary and removed org-checklist module
i am now using the diary on the calendar so that i can have a custom
function, org-diary-last-day-of-month, to get the last day of each month
and i use it as a schedule timestamp. also, i remove org-checklist
module and implemented the only functionality that i wanted to use from
it.