naming changed and added add-hooks and related
i added a couple of functions to core.el to easily add hooks to a list
of major modes or add hooks to a list of hooks. i also am dropping the
naming dotemacs-* in favor of file-* treating each config file as a
"package". i almost finished dropping all the cl-loops.